Remove all the UIKit code from the Dejavu package
